# Move cauchy to Aten(CPU)