--- a/aten/src/ATen/native/Distributions.cpp
+++ b/aten/src/ATen/native/Distributions.cpp
@@ -14,6 +14,7 @@
 #include <ATen/native/DispatchStub.h>
 #include <ATen/native/UnaryOps.h>
 #include <ATen/NamedTensorUtils.h>
+#include <ATen/native/TensorIterator.h>
 
 #include <type_traits>
 #include <functional>
@@ -111,6 +112,10 @@ int64_t sample_poisson(double lambda, at::CPUGenerator* generator) {
 namespace at {
 namespace native {
 
+DEFINE_DISPATCH(bernoulli_mkl_stub);
+DEFINE_DISPATCH(cauchy_stub);
+DEFINE_DISPATCH(multinomial_stub);
+
 Tensor bernoulli(const Tensor& self, Generator* gen) {
   return at::empty_like(self, LEGACY_CONTIGUOUS_MEMORY_FORMAT).bernoulli_(self, gen);
 }
@@ -157,8 +162,6 @@ Tensor& bernoulli_tensor_cpu_(Tensor& self, const Tensor& p_, Generator* gen) {
   return self;
 }
 
-DEFINE_DISPATCH(bernoulli_mkl_stub);
-
 Tensor& bernoulli_scalar_cpu_(Tensor& self, double p, Generator* gen) {
   TORCH_CHECK(0 <= p && p <= 1, "bernoulli_ expects p to be in [0, 1], but got p=", p);
 #if AT_MKL_ENABLED()
@@ -180,6 +183,11 @@ Tensor& bernoulli_scalar_cpu_(Tensor& self, double p, Generator* gen) {
   return self;
 }
 
+Tensor& cauchy_(Tensor& self, double median, double sigma, Generator* gen) {
+  auto iter = TensorIterator::nullary_op(self);
+  cauchy_stub(iter.device_type(), iter, median, sigma, gen);
+  return self;
+}
 
 Tensor _standard_gamma_grad_cpu(const Tensor& self, const Tensor& output) {
   Tensor ret = at::empty(self.sizes(), self.options());
@@ -326,7 +334,5 @@ Tensor multinomial(const Tensor& self, int64_t n_sample, bool with_replacement, 
   native::multinomial_out(result, self, n_sample, with_replacement, gen);
   return result;
 }
-
-DEFINE_DISPATCH(multinomial_stub);
 
 }} // namespace at::native