@kostmo 2020-01-03
## :pill: CircleCI build failures summary and remediations
As of commit 66da0008:


* **1/1** failures introduced in this PR


## Detailed failure analysis

One may explore the probable reasons each build failed interactively [on the Dr. CI website](https://dr.pytorch.org/commit-details.html?sha1=66da0008254c031eaa9befd11e12a0951791dabf).

### :detective: 1 new failure recognized by patterns
The following build failures do not appear to be due to upstream breakage:
#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/4116145) pytorch_xla_linux_xenial_py3_6_clang7_build (1/1)
**Step:** "Build" ([full log](https://dr.pytorch.org/api/view-log-full?build_id=79804533) | [pattern match details](https://dr.pytorch.org/build-details.html?build_id=79804533))
<details>
<summary>
<code>Jan 03 11:28:49 caused by: Connection refused (os error 111)</code>
</summary>

```
Jan 03 11:28:49 +++ eval 'extract_trap_cmd ' 
Jan 03 11:28:49 ++++ extract_trap_cmd 
Jan 03 11:28:49 ++++ printf '%s\n' '' 
Jan 03 11:28:49 +++ printf '%s\n' cleanup 
Jan 03 11:28:49 ++ trap -- ' 
Jan 03 11:28:49 cleanup' EXIT 
Jan 03 11:28:49 ++ which sccache 
Jan 03 11:28:49 ++ sccache --stop-server 
Jan 03 11:28:49 Stopping sccache server... 
Jan 03 11:28:49 error: couldn't connect to server 
Jan 03 11:28:49 caused by: Connection refused (os error 111) 
Jan 03 11:28:49 ++ true 
Jan 03 11:28:49 ++ rm /var/lib/jenkins/sccache_error.log 
Jan 03 11:28:49 rm: cannot remove '/var/lib/jenkins/sccache_error.log': No such file or directory 
Jan 03 11:28:49 ++ true 
Jan 03 11:28:49 ++ SCCACHE_ERROR_LOG=/var/lib/jenkins/sccache_error.log 
Jan 03 11:28:49 ++ SCCACHE_IDLE_TIMEOUT=1200 
Jan 03 11:28:49 ++ RUST_LOG=sccache::server=error 
Jan 03 11:28:49 ++ sccache --start-server 
Jan 03 11:28:49 Starting sccache server... 
Jan 03 11:28:49 ++ sccache --zero-stats 
```

</details>



---

<details><summary>This comment was automatically generated by <a href="https://github.com/kostmo/circleci-failure-tracker/tree/master/docs/from-pull-request-comment">Dr. CI</a> (expand for details).</summary>Follow <a href="https://dr.pytorch.org/admin/comments-opt-out.html">this link to opt-out</a> of these comments for your Pull Requests.

Please report bugs/suggestions on the <a href="https://github.com/kostmo/circleci-failure-tracker/issues">GitHub issue tracker</a>.
</details>